Param set value now throws Exception on error
Control.set_value() method added. This method is similar to
Control.value's setter, but throws an Exception instead of a warning if
the value can't be set. This makes it easier for Python scripts to know
whether the new parameter was successfully set.
Also added a bit extra functionality to the example Python script
(example.py)